requirements.txt
pandas
numpy
pyyaml
progressbar
h5py
seaborn
matplotlib
scikit-learn
tensorflow[and-cuda]
tensorrt
jupyter